requirements.txt
streamlit==1.49.1
google-genai==1.45.0google
generativeai==0.8.5
pandas==2.2.3
pillow==11.3.0